frontend/requirements.txt
streamlit>=1.30.0
requests==2.31.0
pillow==10.1.0
psycopg2-binary==2.9.9
pandas==2.1.3